[ogl,windows] fix resize behavior
This is done in two parts:

1. we now use gl::Viewport to scale the renderable surface to the
window's inner dimensions.

2. due to the way the window events are sent through the system
there is a lag for a large number of frames where the event's
logical_size is catching up to the windows actual size. We work
around this by using the actual size of the window at the time of
the event.