Replace glx with egl
This commit replaces our usage of GLX with EGL. EGL has several
advantages over GLX: its a newer API, its the API used by the Quest, and
it makes it easier to choose a context that supports a specific version
of OpenGL. We use the latter to explicitly select a GLES 3 context. GLES
3 has built-in support for derivatives in shaders, so this will allow us
to remove the polyfills for those functions.