Adjust LiveView widget flick scrolling to behave more natural.
After comparing with native apps in Android, we changed a few things:

We now differentiate between a "short drag", being the one that
defines the swipe movement for the consequent scroll of the list, from
a "long drag".

The latter causes the scroll animation to be stopped and, as a
consequence, the list view is scrolled directly by the drag motion as it
is done by the user.