Fix width and height in Animator
We had an issue using width and height in animations because there are
special cases (Size is an enum and literal numeric values should be
taken as Fixed(..) values).

When attempting to make transitions using numeric values for those
attributes, you got `Apply without file, at ...` errors.

This commit covers missing cases in the related logic in turtle.rs file.