Do not allow to close stack navigation view when it is still opening
It prevents race conditions where the widget ends up in a broken state.
For now we prefer to leave this widget logic simple, by modeling it with
two states: inactive and active. So, when the widget is transitioning to
active, user must wait to be able to close the view (going back to
inactive).