Don't override a user request for a specific nightly version
Without this change running makepad with a locally build version of
rustc_codegen_cranelift doesn't work as rustc_codegen_cranelift needs a
particular nightly rustc version to pair with and trying to load it with
any other version will crash the compiler. cargo-clif handles using the
right nightly version for you, but makepad overrides it before this
change, causing the aformentioned crash.