Limit hover actions emitted by Label and Slider labels
The recent addition into Label widget of hover events actions was
problematic because the widget is not consuming finger events, impeding
other widgets (such as regular views cointaining the labels) to receive
similar events.

This commit adds fine-grained options to Label, Slider and View to
control when the hover actions want to be emitted, leaving the options
turned off for regular labels, which is a better default for now.