`PortalList`: return item's scroll offset along with its item ID/index
Adds a new method: `PortalList::next_visible_item_with_scroll()`.
This is similar to `next_visible_item()`, but it also returns the
item's scroll offset from the top of the `PortalList`'s viewport
in addition to the next visible item's index/ID.

This makes it possible for a user of `PortalList` to track where each visible item has been drawn, which is useful for myriad reasons.

Implementation detail: `next_visible_item()` now calls this new method
for the purposes of code deduplication.