Deliver a `ModalAction::Dismissed` to the inner modal content
This allows the inner modal subview/subwidget to be aware of the fact
that the user has closed the modal *before* the modal actually gets
closed.
This enables myriad ways for the inner modal content to respond to
or otherwise handle the modal being closed without it being a random
surprise.