Tooltip: dismiss on click/press
When displaying a tooltip on mobile devices without hover events,
the tooltip won't be automatically hidden after it has been shown
(e.g., on a long-press event).
Ideally the app dev will choose to hide it upon a different user
interaction, such as clicking elsewhere, but that can be quite
difficult to implement consistently.

This is a very localized change that allows the user to just click
on a tooltip to hide it, which works in all cases that the app dev
inadvertently forgets to otherwise handle.