Redesign `Event::BackPressed` to contain a `handled` bool
The main idea is that a `BackPressed` event should only be handled
*once* by a single widget.
You can use the new `Event::back_pressed()` event to automatically handle this.

This also removes the `Hit::BackPressed` variant, as it doesn't make sense
to have area-based hits for a `BackPressed` event because that event
is not associated with an area-based or focus-based event.

Adjust StackNavigation widget's event handling order to ensure
that the child view's event handler is called first, which allows
child widgets to consume the `BackPressed` event before the
`StackNavigation` widget can consume the `BackPressed` event itself.