Breaking change: allow app logic to mark a `Hit` event as [un]handled
This commit adds a new paradigm that allows the app dev to control
whether the act of handling hits from a given event (when calling `event.hits()`)
will or will not mark that event as handled.

The default is that the `Hit` *does* mark the causing event as handled,
such that it won't be handle-able by other widgets duplicately
(unless of course they enable `capture_overload`).
This commit still preserves that historical behavior in order to not
surprise any users further; however, I do believe that the default
should be that an event is *not* considered to be handled
unless the app logic explicitly marks it as handled.

This currently only applies to four `Hit` variants:
* `Hit::FingerDown`
* `Hit::FingerMove`
* `Hit::FingerHoverIn`
* `Hit::FingerHoverOver`

because those are the four `Hit`s that previously set their event's
`handled` Area unconditionally (internal to the `hits()` API).

Any code that matches on one of those four `Hit` variants will need
to be updated. It's very simple to do so, you can just ignore the extra
`EventHandled` parameter when matching on those variants.
```rust
pub enum Hit {
    ...
    FingerDown(FingerDownEvent, EventHandled),
    FingerMove(FingerMoveEvent, EventHandled),
    FingerHoverIn(FingerHoverEvent, EventHandled),
    FingerHoverOver(FingerHoverEvent, EventHandled),
    ...
```

In addition, all of these Hit handlers within the base `View` widget
now explicitly mark those events as *not* handled, which ensures
they are also able to handled by other widgets.
This makes sense because the base `View` should not "consume" any hits,
since most app devs would not expect that a blank empty space view
would consume a mouse click or something.