Support real OpenHarmony hardware
Fix OpenGL configuration on real OHOS devices, which is different
than what is required for OHOS simulators.

Clarify the instructions for building and signing a real `.hap`
app bundle on a real device.

Ensure that all combinations of Windows, macOS, x86_64, and aarch64
work when building across different hosts and target architectures
for OpenHarmony.