Handle the hover-out action even if a button is disabled
This fixes a subtle bug where a button is enabled when the mouse cursor
hovers-in or over it, but then the button is disabled *before* the
mouse cursor hovers-out away from the button.
This would leave the button's hover state permanently activated
until it is re-enabled, even when the mouse cursor leaves the button.