`Button`: disable long-press detection by default
This change restores the prior functionality of `Button` that allows
it to be treated as "clicked" regardless of how long the button press
was held down.

It introduces a new bool `enable_long_press` which is `false` by default.
If set to `true`, the Button will explicitly handle long presses
(which only occur on mobile platforms that have a native long press event)
and will *not* consider itself to be clicked if a long press occurred
within the button's area.