Make sure emojis are padded correctly.
Outlined glyphs currently have 2 pixels of padding on each side. This is
necessary both to extend the SDF field so that it gets rendered correctly
at the edges and to make sure that glyphs that lie at subpixel boundaries
don't get clipped when their edges don't overlap with a pixel center.

This commit adds the same amount of padding to rasterized glyphs (i.e.
emojis), so that these won't clipped either.

One consequence of this is that we now add manual padding for both outlined
and rasterized glyphs, so the implicit padding in the atlas (which was
there to make sure glyph textures don't bleed into each other) is no longer
necessary.