Ensure that LongPress (LongClick) on Android properly tracks touch movement (#820)
* Ensure that LongPress (LongClick) on Android properly uses touch slop

Previously, we assumed that multiple touch action events would occur
before a LongClick, but that is not necessarily true.
It is possible to just have one down touch immediately followed by a
LongClick, so we now account for that.

This also fixes the tracking of touch event locations such that
stale values aren't accidentally used when calculating if a finger movement
exceeded the allowable touch slop for considering a touch as a long press.
(Rust would've caught that... thanks Java)

* remove excess log stmt